pyproject.toml
[project]
name = "convergent"
version = "0.1.0-phase0"
description = "Strategic Tax Optimization Suite — deterministic rules engine for closely-held business owners"
readme = "README.md"
requires-python = ">=3.12,<3.13"
authors = [{ name = "Convergent" }]
license = { text = "Proprietary" }

# Dependencies are pinned to exact versions. Phase 0 target — any upgrade must be
# reviewed and re-tested against the full golden-scenario suite.
dependencies = [
    # Numerics & data
    "numpy==2.1.3",
    "scipy==1.14.1",
    "pandas==2.2.3",
    "pydantic==2.9.2",
    "pulp==2.9.0",

    # Persistence
    "sqlalchemy==2.0.36",
    "pysqlcipher3==1.2.0; sys_platform == 'win32'",

    # UI
    "nicegui==2.7.0",
    "plotly==5.24.1",
    "pywebview==5.3.2",

    # Anthropic SDK (Authority Layer only)
    "anthropic==0.39.0",

    # OCR / document ingestion
    "pytesseract==0.3.13",
    "pdfplumber==0.11.4",
    "camelot-py[cv]==0.11.0",
    "pillow==11.0.0",
    "pillow-heif==0.20.0",

    # Document I/O
    "python-docx==1.1.2",
    "openpyxl==3.1.5",
    "reportlab==4.2.5",
    "docx2pdf==0.1.8; sys_platform == 'win32'",
    "jinja2==3.1.4",

    # Secrets (Windows Credential Manager)
    "keyring==25.5.0",

    # Windows-only utilities
    "pywin32==308; sys_platform == 'win32'",

    # HTTP / scraping for Statutory Mining
    "httpx==0.27.2",
    "beautifulsoup4==4.12.3",
    "lxml==5.3.0",

    # Crypto for portable export
    "cryptography==43.0.3",
    "argon2-cffi==23.1.0",

    # Misc
    "pyyaml==6.0.2",
    "structlog==24.4.0",
    "tenacity==9.0.0",
    "mermaid-py==0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "playwright==1.48.0",
    "pytest-playwright==0.5.2",
    "ruff==0.7.3",
    "mypy==1.13.0",
    "pyinstaller==6.11.1",
]

[project.scripts]
convergent = "convergent.__main__:main"
convergent-bootstrap = "convergent.authority_layer.statutory_mining.bootstrap:run"
convergent-mine = "convergent.authority_layer.statutory_mining.scheduler:run_once"

[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["convergent*"]

[tool.setuptools.package-data]
convergent = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers --strict-config"
markers = [
    "golden: golden scenario end-to-end tests (slow)",
    "authority: authority layer regression tests (requires ANTHROPIC_API_KEY)",
    "ui: UI smoke tests (requires Playwright browsers)",
    "windows: Windows-only tests",
]
